requirements.txt
pika >= 1.2.0
redis >= 4.1.1
boto3 >= 1.21.42
python-dotenv